// packages/sdk/src/generated/types/memberWithVerifyArgs.ts
/**
 * This code was AUTOGENERATED using the codama library.
 * Please DO NOT EDIT THIS FILE, instead use visitors
 * to add features, then rerun codama to update it.
 *
 * @see https://github.com/codama-idl/codama
 */

import {
  combineCodec,
  getOptionDecoder,
  getOptionEncoder,
  getStructDecoder,
  getStructEncoder,
  type Codec,
  type Decoder,
  type Encoder,
  type Option,
  type OptionOrNullable,
} from '@solana/kit';
import {
  getMemberDecoder,
  getMemberEncoder,
  getSecp256r1VerifyArgsDecoder,
  getSecp256r1VerifyArgsEncoder,
  type Member,
  type MemberArgs,
  type Secp256r1VerifyArgs,
  type Secp256r1VerifyArgsArgs,
} from '.';

export type MemberWithVerifyArgs = {
  data: Member;
  verifyArgs: Option<Secp256r1VerifyArgs>;
};

export type MemberWithVerifyArgsArgs = {
  data: MemberArgs;
  verifyArgs: OptionOrNullable<Secp256r1VerifyArgsArgs>;
};

export function getMemberWithVerifyArgsEncoder(): Encoder<MemberWithVerifyArgsArgs> {
  return getStructEncoder([
    ['data', getMemberEncoder()],
    ['verifyArgs', getOptionEncoder(getSecp256r1VerifyArgsEncoder())],
  ]);
}

export function getMemberWithVerifyArgsDecoder(): Decoder<MemberWithVerifyArgs> {
  return getStructDecoder([
    ['data', getMemberDecoder()],
    ['verifyArgs', getOptionDecoder(getSecp256r1VerifyArgsDecoder())],
  ]);
}

export function getMemberWithVerifyArgsCodec(): Codec<
  MemberWithVerifyArgsArgs,
  MemberWithVerifyArgs
> {
  return combineCodec(
    getMemberWithVerifyArgsEncoder(),
    getMemberWithVerifyArgsDecoder()
  );
}